requirements.txt
langchain
langchain-community
langchain-weaviate
langchain-ollama
weaviate-client
fastapi
uvicorn
pypdf
tiktoken
python-multipart
ragas
datasets
pandas